Add new 'FindStream' test in repo_test.go
- Imported "time" package.
- Created a new struct FindStreamModel for testing purpose.
- Added a new test function "TestRepository_FindStream" which includes the code for connecting MongoDB, inserting data, reading data from the database and comparing the read data with the inserted data.

The new function essentially tests the functionality of the FindStream method in the repository. It tests the ability to find and retrieve specific streams of data from the MongoDB. The function also tests error handling and disconnection from the MongoDB. The test is conducted using mock data entered into the MongoDB and then retrieved using the FindStream method.